render playlist page
