+++ src/lib/stores.js
@@ -1,5 +1,6 @@
 import { writable } from 'svelte/store';
 
+export const playlistNameStore = writable('');
 export const searchResults = writable([]);
 export const notFoundSearchResults = writable([]);
